requirements.txt
sqlalchemy
psycopg2-binary
playwright
pandas
apscheduler
cryptography
python-dotenv
fastapi
uvicorn
